refactor: update import paths for RoutingProtocolService in job and cron-job services
